client/requirements.txt
requests
python-csv
regex
webvtt-py